requirements.txt
alabaster==1.0.0
babel==2.17.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
docutils==0.21.2
h11==0.14.0
idna==3.10
imagesize==1.4.1
importlib_metadata==8.6.1
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==24.2
Pygments==2.19.1
pytz==2025.1
requests==2.32.3
roman-numerals-py==3.1.0
setuptools==76.0.0
snowballstemmer==2.2.0
Sphinx==8.2.3
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
zipp==3.21.0
python-dotenv==1.0.1
langchain==0.3.20
langchain-openai==0.3.8
langchain-community==0.3.19
faiss-cpu==1.10.0
pillow==11.1.0
pdf2image==1.17.0
Office365-REST-Python-Client==2.5.14
setuptools==76.0.0
fastapi==0.115.11
psycopg2-binary==2.9.10
aiofiles==24.1.0
python-multipart==0.0.20
itsdangerous==2.2.0
aiofiles
pydantic[email]
langgraph==0.4.7
-e .
